% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeoMeanNegProbes.R
\name{checkGeoMeanNegProbes}
\alias{checkGeoMeanNegProbes}
\title{Assessment of NegGeoMean}
\usage{
checkGeoMeanNegProbes(eSet, minNegGeoMean = 1)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet object.}

\item{minNegGeoMean}{Threshold to define the minimum required NegGeoMean. Default is 1.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
The geometric mean of the negative control probes (NegGeoMean) is used to distinguish signal from noise and determines the quantifiable limit of gene expression. Segments with a very low NegGeoMean often exhibit low gene detection rates.
}
\examples{
checkGeoMeanNegProbes(eSet)
}
