% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNTC.R
\name{checkNTC}
\alias{checkNTC}
\title{Checks the performance of NTCs}
\usage{
checkNTC(eSet = eSet, Raw = "Raw")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{Raw}{Name of the column containing the number of raw sequencing reads.}
}
\value{
Nothing.
}
\description{
Checks the performance of the No Template Control (NTC) in a GeoMx DSP experiment.
}
\examples{
checkNTC(eSet)
}
