% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNTCcounts.R
\name{checkNTCcounts}
\alias{checkNTCcounts}
\title{Check NTC count uniformity}
\usage{
checkNTCcounts(eSet, Raw = "Raw")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{Raw}{Name of the column containing the number of raw sequencing reads.}
}
\value{
A ggplot2 object.
}
\description{
Plots a histogram of counts for each NTC to assess uniformity.
}
\examples{
checkNTCcounts(eSet)
}
