% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNuclei.R
\name{checkNuclei}
\alias{checkNuclei}
\title{Assessment of the relationship between Nuclei and Area}
\usage{
checkNuclei(eSet = eSet, removeZero = TRUE, Nuclei = "Nuclei", Area = "Area")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{removeZero}{If TRUE, will discard segments where either the number of nuclei or the area size is 0.}

\item{Nuclei}{The column containing the nuclei counts.}

\item{Area}{The column containing the area size.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Assesses the relationship between the number of nuclei and the area size across with experiment. Segments which deviate from this relationship are flagged.
}
\examples{
checkNuclei(eSet)
}
