% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRawSequencing.R
\name{checkRawSequencing}
\alias{checkRawSequencing}
\title{Relationship between segment size and number of sequencing reads}
\usage{
checkRawSequencing(eSet = eSet, removeZero = TRUE, Area = "Area", Raw = "Raw")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{removeZero}{If TRUE, will discard segments where either the number of nuclei or the area size is 0.}

\item{Area}{The column containing the area size.}

\item{Raw}{The column containing the number of raw sequencing reads.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
The relationship between area size and the number of sequencing reads across the experiment is assessed and any segments which deviating from this relationship are flagged.
}
\examples{
checkRawSequencing(eSet)
}
