% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSaturation.R
\name{checkSaturation}
\alias{checkSaturation}
\title{Assessment of sequencing saturation}
\usage{
checkSaturation(
  eSet = eSet,
  minSaturation = 50,
  DeduplicatedReads = "DeduplicatedReads",
  Aligned = "Aligned"
)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet()}

\item{minSaturation}{Segments below this threshold are flagged. Default is 50.}

\item{DeduplicatedReads}{The column containing the number of deduplicated reads.}

\item{Aligned}{The column containing the number of aligned reads.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Calculates the sequencing saturation for each sample and flags those with very low sequencing saturation.
}
\examples{
checkSaturation(eSet)
}
