% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPseudobulk.R
\name{createPseudobulk}
\alias{createPseudobulk}
\title{Generates pseudobulk data}
\usage{
createPseudobulk(eSet, id.vars)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet object.}

\item{id.vars}{Variable corresponding to the column name used for melting.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Creates a pseudobulk NanoStringGeoMxSet object by aggregating gene counts from all segments belonging to the same sample.
}
\examples{
createPseudobulk(eSet, id.vars = "SampleID")
}
