% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSTlist.R
\name{createSTlist}
\alias{createSTlist}
\title{Create an STlist object}
\usage{
createSTlist(eSet, id.vars)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet object.}

\item{id.vars}{Variable corresponding to the column name used for melting.}
}
\value{
An STlist object.
}
\description{
Creates an STlist object for downstream analysis with \code{spatialGE}
}
\examples{
createSTlist(eSet, id.vars = "SampleID")
}
