% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dustR-package.R
\docType{package}
\name{dustR-package}
\alias{dustR}
\alias{dustR-package}
\title{dustR: Tidying GeoMx Whole Transcriptome Atlas spatial transcriptomics data in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Designed to optimise quality assurance of GeoMx Whole Transcriptome Atlas (WTA) spatial transcriptomics data by decreasing segment loss and increasing gene detection rates.
}
\author{
\strong{Maintainer}: Jelmar Quist \email{jelmar.quist@kcl.ac.uk}

Other contributors:
\itemize{
  \item Isobelle Wall \email{isobelle.wall@kcl.ac.uk} [contributor]
}

}
\keyword{internal}
