% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNegativeProbes.R
\name{filterNegativeProbes}
\alias{filterNegativeProbes}
\title{Filtering of negative control probes}
\usage{
filterNegativeProbes(eSet = eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Inspects the relationship between the number of sequencing reads and the abundance of each negative control probe.
}
\examples{
filterNegativeProbes(eSet)
}
