% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSegments.R
\name{filterSegments}
\alias{filterSegments}
\title{Filtering of segments}
\usage{
filterSegments(eSet = eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Function that removes all segments flagged during the quality control process.
}
\examples{
filterSegments(eSet)
}
