% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlope.R
\name{getSlope}
\alias{getSlope}
\title{Estimate slope}
\usage{
getSlope(NegativeProbes_data)
}
\arguments{
\item{NegativeProbes_data}{List of negative control probes containing DCC, Raw and Count
'}
}
\value{
Estimated slope.
}
\description{
Performs a glm() and calculates the slope.
}
\examples{
getSlope(NegativeProbes_data)
}
