% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFilterSegments.R
\name{plotFilterSegments}
\alias{plotFilterSegments}
\title{Plot of segments of poor quality}
\usage{
plotFilterSegments(eSet = eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}
}
\value{
A ggplot2 object.
}
\description{
A summary plot depicting the various filtering criteria. Some of these criteria are likely interconnected.
}
\examples{
plotFilterSegments(eSet)
}
