% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeoMeanNegProbes.R
\name{plotGeoMeanNegProbes}
\alias{plotGeoMeanNegProbes}
\title{Plot geometric mean of the negative control probes.}
\usage{
plotGeoMeanNegProbes(
  eSet,
  minNegGeoMean = 1,
  NegGeoMean = "NegGeoMean",
  Segment = "Segment"
)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{minNegGeoMean}{Segments below this threshold are flagged. Default is 1.}

\item{NegGeoMean}{The column reporting the geometric mean of the negative control probes.}

\item{Segment}{Name of column containing segment information (e.g. PanCK+, CD45, ...).}
}
\value{
A ggplot2 object.
}
\description{
A barplot of the geometric mean of the negative control probes by segment.
}
\examples{
plotGeoMeanNegProbes(eSet)
}
