% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNegativeProbes.R
\name{plotNegativeProbes}
\alias{plotNegativeProbes}
\title{Plot  of negative control probe performance.}
\usage{
plotNegativeProbes(eSet = eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}
}
\value{
A ggplot2 object.
}
\description{
Plots the relationship between the number of sequencing reads and the abundance of each negative control probe.
}
\examples{
plotNegativeProbes(eSet)
}
