% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNTC.R
\name{plotNTC}
\alias{plotNTC}
\title{Plot the number of raw sequencing read for each NTC}
\usage{
plotNTC(eSet = eSet, Raw = "Raw")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{Raw}{Name of the column containing the number of raw sequencing reads.}
}
\value{
A ggplot2 object.
}
\description{
Plots the number of sequencing reads detected in each No Template Control (NTC).
}
\examples{
plotNTC(eSet)
}
