% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawSequencing.R
\name{plotRawSequencing}
\alias{plotRawSequencing}
\title{Plot Area and Sequencing Reads}
\usage{
plotRawSequencing(eSet = eSet, Area = "Area", Raw = "Raw")
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{Area}{The column containing the area size.}

\item{Raw}{The column containing the number of raw sequencing reads.}
}
\value{
A ggplot2 object.
}
\description{
Plots the relationship between the segment area and the number of sequencing reads. Poor performing segments (from checkSequencing()), are indicated in the plot.
}
\examples{
plotRawSequencing(eSet)
}
