% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSaturation.R
\name{plotSaturation}
\alias{plotSaturation}
\title{Barplot of sequencing saturation}
\usage{
plotSaturation(eSet, minSaturation = 50)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{minSaturation}{Segments below this threshold are flagged. Default is 50.}
}
\value{
A ggplot2 object.
}
\description{
A barplot of sequencing saturation across all segments.
}
\examples{
plotSaturation(eSet)
}
