% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNegativeProbe.R
\name{setNegativeProbe}
\alias{setNegativeProbe}
\title{Flags a set of negative control probes}
\usage{
setNegativeProbe(negativeProbes_data)
}
\arguments{
\item{negativeProbes_data}{A data.frame of RTS and Slope}
}
\value{
A flag ("PASS" or "FAIL") for each negative control probe.
}
\description{
Identifies negative control probes of poor performance by assessing the relationship between its counts and the number of raw sequencing reads.
}
\examples{
setNegativeProbe(negativeProbes_data)
}
