% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeoMeanNegProbes.R
\name{checkGeoMeanNegProbes}
\alias{checkGeoMeanNegProbes}
\title{Check Negative Control Probes}
\usage{
checkGeoMeanNegProbes(
  eSet = eSet,
  minNegGeoMean = 1,
  NegGeoMean = "NegGeoMean"
)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}

\item{minNegGeoMean}{Segments below this threshold are flagged. Default is 1.}

\item{NegGeoMean}{The column reporting the geometric mean of the negative control probes.}
}
\value{
A readNanoStringGeoMxSet object.
}
\description{
Geometric mean of the negative control probes determines the limit of quantification and subsequentially the gene detection rate. Segments with a low geometric mean are flagged.
}
\examples{
checkGeoMeanNegProbes(eSet)
}
