% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNTC.R
\name{checkNTC}
\alias{checkNTC}
\title{No Template Control assessment}
\usage{
checkNTC(eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet object.}
}
\value{
Nothing.
}
\description{
A No Template Control (NTC) is included in each sequencing plate (well A01) to detect contamination of the sequencing library. NTCs with more than 1,000 sequencing reads could indicate contamination and should be examined further.
}
\examples{
checkNTC(eSet)
}
