% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNTCcounts.R
\name{checkNTCcounts}
\alias{checkNTCcounts}
\title{No Template Control count uniformity}
\usage{
checkNTCcounts(eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet object.}
}
\value{
A ggplot2 object.
}
\description{
In No Template Controls (NTCs) with more than 1,000 sequencing reads, counts are expected to be uniformly distributed (i.e. detected at a similar abundances). This is assessed by comparing the gene count distribution of each NTC with a uniform distribution of comparable parameters using a Kolmogorov–Smirnov test.
}
\examples{
checkNTCcounts(eSet)
}
