% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNuclei.R
\name{plotNuclei}
\alias{plotNuclei}
\title{Plot Nuclei versus Area}
\usage{
plotNuclei(eSet = eSet)
}
\arguments{
\item{eSet}{A NanoStringGeoMxSet acquired via readNanoStringGeoMxSet().}
}
\value{
A ggplot2 object.
}
\description{
Plots the relationship between the number of nuclei and the area size to identify poor performing segments.
}
\examples{
plotNuclei(eSet)
}
